style dictionary config refactor and sizing json reorganization
